Use long render-list frame tracking

+++ src/main/java/me/cortex/voxy/client/core/rendering/section/backend/vulkanberyl/VulkanBerylViewportRenderList.java
@@ -19,7 +19,7 @@ public final class VulkanBerylViewportRenderList implements ViewportRenderList {
     private final int maxEntryCount;
     private final Buffer buffer;
     private int lastVisibleCount = -1;
-    private int lastVisibleFrameId = -1;
+    private long lastVisibleFrameId = -1L;
     private boolean freed;
 
     public VulkanBerylViewportRenderList() {
@@ -59,15 +59,15 @@ public final class VulkanBerylViewportRenderList implements ViewportRenderList {
         return this.lastVisibleCount;
     }
 
-    public int getLastVisibleFrameId() {
+    public long getLastVisibleFrameId() {
         return this.lastVisibleFrameId;
     }
 
     void setLastVisibleCount(int lastVisibleCount) {
-        setLastVisibleCount(lastVisibleCount, -1);
+        setLastVisibleCount(lastVisibleCount, -1L);
     }
 
-    void setLastVisibleCount(int lastVisibleCount, int frameId) {
+    void setLastVisibleCount(int lastVisibleCount, long frameId) {
         this.lastVisibleCount = lastVisibleCount;
         this.lastVisibleFrameId = frameId;
     }
